This push brings some improvements to the stats package with more flexibility in Percentil and a new MovingWindowDelta Stat.
Also in this push:

build:
  + fixup forked junit vm to inherit the ANT_OPTS jvm flags

pants:
  + fix ide commands to work with scala leaf targets
  + allow java_library and scala_library to float between src and tests
  + allow python_antlr_library to float anywhere in the source tree

+++ src/java/com/twitter/common/stats/RequestStats.java
@@ -31,7 +31,7 @@ public class RequestStats implements RequestTimer {
 
   private static final float DEFAULT_SAMPLE_PERCENT = 10;
   // TODO(William Farner): Add fractional percentiles, requires support in Percentile.
-  private static final int[] DEFAULT_PERCENTILES = {10, 50, 90, 99};
+  private static final double[] DEFAULT_PERCENTILES = {10, 50, 90, 99};
 
   private final SlidingStats requests;
   private final Percentile<Long> percentile;
